requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug
python-dotenv
pandas
matplotlib
pyttsx3
openai-whisper
langchain
langchain-community
langchain-core
google-genai
streamlit-audiorecorder
flask_babel
reportlab
flask_restx
flask_cors
flask_migrate
psycopg2-binary
redis
Flask-Session
Flask-Caching
celery
